requirements.txt
bcrypt==4.2.0
cffi==1.17.1
cryptography==43.0.3
paramiko==3.5.0
pycparser==2.22
PyNaCl==1.5.0
python-decouple==3.8
